installkernel: rm before mv to prevent error on hardlinks
If the `.old` file is hard-linked to the file that is about to moved onto it,
then the `mv` will fail:
```
$ mv vmlinuz-6.1.22-gentoo-dist vmlinuz-6.1.22-gentoo-dist.old
mv: 'vmlinuz-6.1.22-gentoo-dist' and 'vmlinuz-6.1.22-gentoo-dist.old' are the same file
```

We could conditionally test if it's hardlinked, but since we are moving anyway,
just delete the `.old` file before the `mv`.

Testing for being hardlinked requires either non-POSIX `test $F1 -ef $F2` or
GNU find `-samefile`, whereas the `rm ; mv` solution will work on POSIX shell.

Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>