installkernel-9999.ebuild: depend on sd-stub when dracut and uki
Systemd-stub is required when building an UKI (with ukify or dracut),
the systemd-stub is installed when the boot flag on systemd(-utils) is
enabled.

This fixes a common issue new UKI users run into when building UKIs with
dracut. Ukify already pulls in the stub via the systemd(-utils) ebuild, but
lets also make this explicit anyway.

Note we don't want to depend on sd-stub conditionally to use uki only
because when generic-uki is used then no uki will be generated locally
and therefore sd-stub is not required but the uki flag should still be
enabled in this case.

Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>