installkernel-9999.ebuild: switch priority of grub and sd-boot
UKI users need sd-stub, provided by the boot flag on systemd(-utils). When this
flag is enabled but the systemd-boot flag is not enabled on installkernel,
users will see a warning message about this. This is helpful for systemd-boot
users, but users who wish to boot with UKI and grub may get confused and enable
the systemd-boot flag on installkernel. By switching the order around here, we
avoid that this confusion leads to boot failure when the kernels are installed
in the bls layout that grub does not understand.

https://forums.gentoo.org/viewtopic-t-1167496.html

Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>