run_tests.sh: adjust cases with systemd-boot and grub
This is expected since we adjusted the priority of sd-boot and grub. Will not
effect real users since enabling both sd-boot and grub is invalid anyway.

The changing of priority "fixes" the case where an user wants to boot
UKIs with grub, therefore needs to enable boot on systemd(-utils) to get
sd-stub. But then is warned that the systemd-boot flag is not enabled,
configuration is "correct" though the warning message may cause the user to
enable the systemd-boot flag which would then cause the system to break.
Switching priority fixes this situation (though of course enabling sd-boot
and grub at the same time is weird and invalid regardless)

Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>