*: use install instead of cp when installing to possible FAT location
cp --preserve was a mistake. The proper solution to the problem
raised in projg2/installkernel-gentoo/pull/32 is to umask the
entire partition in fstab. This is also recommended by systemd, and
tools such as bootctl will complain if this is not the case. Instead
of using cp, we use install, and we use it in the exact same way that
the hooks that ship with systemd use it for consistency. Blindly
preserving the original mode was never a good idea because we cannot
know what the original mode is. Dracut may explicitly umask it to 077
but this is no guarantee for other initramfs generators. And also, if
the original file was not owned by root we will create all sorts of
weirdness trying to preserve that on a FAT file system. So partially
revert 04f43c7c06c30bf5a3ffd723c9b7bd72879552c0 and
aad29bd40ae0de81760b300840dc28b851683a5b and just do what systemd does
instead. We can keep the --preserve stuff when copying in the same
directory (such as backing up grub.cfg), or when writing to the
staging area.

See-also: https://github.com/projg2/installkernel-gentoo/pull/32
Closes: https://github.com/projg2/installkernel-gentoo/issues/34
Signed-off-by: Nowa Ammerlaan <nowa@gentoo.org>