*: drop silly "please run as root" message
The only reason that is in all these hooks is because that is how
I found the original one they are all forked from. But if we want to
do this we should really just should check it in the main script, but
lets not do that for a) consistency with systemd kernel-install and
b) because running this as a regular user is actually useful if you
want to test your dracut/ukify configuration without actually doing
anything.

Signed-off-by: Nowa Ammerlaan <nowa@gentoo.org>